requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
PyMuPDF 
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
pydantic == 2.7.4
langchain_chroma
